Disable dock audio settings if the dock doesn't have Bluetooth.
We should really have a separate string here to indicate
this case, but it's pretty late for that sort of thing, so I
left it at the generic "configure dock audio".

We might also want to dim this setting rather than popping
up a "Must dock phone first" dialog if the phone is not
docked.

Bug: 2469862
Change-Id: I4c61f5a50baac55881f5a21e523c370c456f0be8