Fix crash for tether with quotes in SSID
Multiple quotes cause havoc with netd interaction right now.
The solution is to revert the changes for WifiConfiguration on
Access Point side done in Change I0f1e508b and remove all quotes
in the SSID.

Essentially a low risk one line change on top of the revert.

We dont expose API for WifiConfiguration on AP side. This solution
essentially means we treat SSID in WifiConfiguration on AP side
differently. When we do expose things, we should make things consistent.

Change-Id: I0a7e5074b83ace743fc843ee61c1ff86cde8267f