Change BT pairing UI PIN dialog string for clarity.
When pairing with a Bluetooth keyboard, the user may not realize
that the PIN should be entered on the keyboard as well as the device.
Added "You may need to enter the same PIN on the Bluetooth device."
to pairing dialog string. Also, remove unused Bluetooth device info
debug strings.

Bug: 3270429
Change-Id: I6d7fa70cf7b12c1feafb8347f356c568b10ae646