Settings may show enabled accessiiblity services after restore even if accessibility is off.
Now the state of accessibility is determined from the number of running, i.e. enabled and
installed, accessibility services - more than zero access is on else it is off. After a
restore the device may have some enabled and installed services but disabled accessibility
setting. In such case settings should respect the restored values and the services should
be displayed as disabled initially and then enabling one of them will enable accessibility.

bug:5367404

Change-Id: If3756aba3f59e282000247e99f87ebff2b64c9fd