Call setLanguage() when the TTS locale is changed in settings.
This is required to fix the case where an app has instantiated
and is using a TextToSpeech instance, and the user switches to
settings and changes the language for that instance. If no
app was using TTS, then onCreate() would have be called when
TTS is used next, and the correct locale will be loaded.

bug:5823836
Change-Id: I62adf3f7f0f14b67450ba08b8743e0d31e615a7e