Remove "Searching/Tap To Pair" text from Bluetooth Settings.
The second text field next to "Available devices" in the Bluetooth
Settings device list creates a cluttered heading that wraps to multiple
lines in some locales. It's unnecessary because there's a graphical
progress spinner when searching for devices, and "Touch to pair" is
easily discoverable. Remove the second text field and the "Searching"
and "Touch to pair" string resources.

Bug: 5496221
Change-Id: I7b71222415152014426cf54fc112fe5379d350dc