Data usage limit label font size fixed
Bug 6326750

The font size around the text was removed when its content is replaced.
This is WAI since the spans created from <font> html tags have a
SPAN_EXCLUSIVE_EXCLUSIVE flag. These are removed when their length
becomes 0. The bug came from a recent enforcement of this constraint.

The fix is to change the flags of these AbsoluteSizeSpan spans.

This CL also changes the code to avoid the creation of the int[] array.

Change-Id: Ib58270bb9d1cf0c7609ba1e36d6c0cca841300db