Settings: Fixed German translation for turning WLAN/Bluetooth on
Prior to this patch, the question "WLAN aktivieren?" would show
after you turned it on, and while it is loading; this patch
merely addresses this issue and changes it to "WLAN wird aktiviert",
which is consistent with the message showing up when turning it off.

Likewise for turning Bluetooth on.

Change-Id: Ic91143836a85751c39839daae720ac03b174bfde