If external storage is emulated, don't show it as separate entries.
The installed app details UI is really complicted for the two-partition
internal and external storage case...  however, in our lovely new world
of emulated external storage, there is no reason for this, since the
external storage is using the same partition as internal storage.  So
in this case, don't show external storage as a separate item, just
combine it into the base app size and data size values.

Change-Id: Iea0e78f2d532ad50e20fef98740c93b2e897890a