Moved KEY_VIBRATE to NEED_VOICE_CAPABILITY
There is an array (NEED_VOICE_CAPABILITY) with all keys that
should not be present in a device that does not have voice
capability.

The key KEY_VIBRATE was not part of this array, it was handled
separately in the code.

Moved KEY_VIBRATE to NEED_VOICE_CAPABILITY and remove the
special treatement.

Change-Id: I96aa066cec373ccbfe89c46b530feb484f6f8164