Make wifi dialog persist after orientation change
Wifi "cancel/forget/connect" pop-up window does not persist after orientation
changes. Specifically, after orientation changes, the connect button disappears.

This is because dialog information is not saved properly. When a saved wifi
config is found in the scan results, this entry is BOTH a wifi config AND a scan
result. Previous implementation has 2 issues:

1. Scan result part isn't saved. That's why the connect button after orientation
   changes, as UI didn't see a scan result section and thought it's not in range.

2. When both wifi config and scan result are saved, they are not restored
   correctly. Restoring wifi config will reset mRssi field in scan result part,
   while restoring scan result will reset netowrkId in wifi config part.

Both issuses are fixed.

bug: 9053186
Change-Id: Ic3ee7b5931c91d4b2f5c723fb8718cf752463109