Battery: report calculated power, tweak signal strength colors.
BatteryHistoryHelper now computes the total power to be the total
including the unaccounted, so we don't go > 100%.  However for the
unaccounted/over-counted details, we want to report the actual
computed amount.  Use a new API to do this.

Also tweak the signal strength colors so that the "no signal" state
is stronger, by being reddish instead of bright yellow.

Change-Id: I6af9fbcce8c3324883db0184f7ac215cb41b6230