Zen mode settings.
Step 1: Surface zen mode value under sound settings.

This is in keeping with the principle that quick settings are
accelerators (not replacements) for settings in the Settings app.

Adding configuration to "Limited interruptions" will be done as
a followup, and will serve as the target for the configuration
icon in SystemUI.

Change-Id: Ia2c7351454b17df3e27926ff593eebff284ebef8