Add explicit "use system language" TTS language.
Previously, if user set the TTS engine default language to the system
language, Settings app would save its value as empty string - which
is interpreted as "use system language", and makes it follow the change
of the device locale. This change introduces explicit "Use system language"
as an entry in the TTS default engine picker. This makes it consistent
with other pickers, like ones for keyboard and spellchecker.

Change-Id: Ibb4189be30bb9691ef2bd10e7fff883318eceae7