On factory reset, wipe the stored account challenges
Each time a user adds an account, a challenge gets
stored to the persistent data block service.

If a user factory resets *from Settings*, this is
considered a "secure wipe" and thus we must wipe
the account challenges as well.

No other factory reset mechanism must be permitted
to wipe this data.

Bug: 14288780
Change-Id: Ibe48ccae2d7b5f8d736717a27b7c4f44bed0fba7