[Setup] Theme ChooseLockGeneric for setup wizard
This is step one of theming the set-up lock screen flow to match the
setup wizard theme. This shows the general approach of creating a
subclass of both the activity and fragment and overriding methods to
achieve the desired behavior for setup.

ag/594000 is a much more comprehensive change for what the final
change will look like.

Bug: 18482708
Change-Id: Idff34937f39f46a0c488df2cae4c46155b80cab7