[FRP] Remove None and Swipe from screen lock options
Removed "None" and "Swipe" from screen lock options in the choosing
lock screen during setup wizard, because at this step, the user has
already chosen to pick a screen lock, hence it's unnecessory to show
"none" and "swipe".

Implemented this by overriding the DisableUnusablePreferences in
SetupChooseLockGenericFragment to disable the insecure lock methods
and adding a flag to hide disabled methods for setup wizard.

Bug: 18631416
Change-Id: I3feccf591da25bd6eadcc03c75b28cfc1d069db7