[WifiSetup] Don't show menu during setup
Hides the menu during Wi-Fi setup -- we removed the menu button from
the UI but the menu can still be started with a hardware menu button.
This change removes that ability as well (during setup).

If you want to test on a device without a hardware menu button, you
can use the command
    adb shell input keyevent 82

Change-Id: Ibe3cc679353a83c995e24680c3e7dc1575d32fe0