Remove grey Wi-Fi icon from Setup Wizard
This removes the grey Wi-Fi icon from the empty-view in the Wi-Fi
list of Setup Wizard. Also updated the style so it looks like the
description text in other setup screens.
Also fixed the view recycling by not calling getRootAdapter in
onActivityCreated, since that will cause the Preference framework
to not cache the AccessPoint preference.

Bug: 17575719
Bug: 16518752
Change-Id: I06cbc88db1c04b29a9ce8221596265effc39b90b