Implement new Manage Apps list
Manage apps now only has one list.  When entered from the dashboard
it is not sortable or filterable, and shows a combination of downloaded
apps and anything that shows up in launcher.  When 'all apps' is
entered through advanced app settings, the list shows all apps and
can be filtered and sorted.

Bug: 19443900
Change-Id: I555ee56417c330424d458750b6f81d3bd543f66d