[WifiSetup] Update Wi-Fi icon colors
For the new blue-themed setup wizard, update the colors of the Wi-Fi
icons. This reuses the abandoned wifi_signal_colored XML, update it
to reference the new material wifi_signal_teal icon. Only the shape
of wifi_signal_teal is used. Tint is applied to the icon so that the
actual color is defined by wifi_signal_color attr in the theme.

Bug: 19729981
Change-Id: I59a8c137086f2e80eb7e10880626098243b14d13