[WifiSettings] Align text in Wi-Fi dialog
The themes for Wi-Fi dialog in sw600dp is outdated and causing the
labels and text fields to be misaligned. This CL removes all the
related styles except for the one in value/styles.xml, so that all
screens use the same styles (the items and values in the styles were
all the same except for sw600dp which this CL fixes).

Bug: 20663443
Change-Id: I59a079c1a4cc1ae3e99d863cb24816e609e98109