Show storage slow dialog allowing state loss.
We're not picky about losing state here, since the adoption actually
finished under the hood and we're just warning the user that it's
slow.

Without this change, we could end up throwing and re-adopting the
same device twice, which causes issues on sprout-style devices that
only migrate primary storage once.

Bug: 22867250
Change-Id: I1491679ab802f19fe25b82e9b452ca3dabdfec0d