Fix MMS delivery issue by permitting HTTP from Settings app.
MMS delivery may fail on carriers which use cleartext HTTP for MMS,
due to an issue introduced in Android M. The cause is that the
platform-provided HTTP stack refuses cleartext HTTP from the MMS
service process because other packages (Settings app,
TelephonyProvider) loaded into the com.android.phone process declare
a policy that cleartext network traffic is not permitted. The issue
might be non-deterministic because the policy depends on the order in
which packages are loaded into the process.

From the perspective of platform-provided HTTP stacks, whether
cleartext HTTP traffic is permitted is a per-process setting. It is
set by the framework based on the policy of the package due to which
the process is started. Further packages loaded into the process do
not affect the per-process policy. This means all packages which may
be loaded into the same process must use the same cleartext network
traffic policy.

This fix relaxes the network security policy of the Settings app to
permit cleartext network traffic, so that all packages loaded into
the com.android.phone process uses the same policy.

Bug: 24222050
Change-Id: Id6d0018b30782c349963d92799d74266e5001729