Add fallback encryption-aware home screen.
When the home screen selected by the user isn't encryption aware, we
still need to put something on the ActivityStack.  For now, let's use
an empty activity that knows how to dismiss itself when the
credential-encrypted storage is unlocked; that's enough for the
system to re-resolve the home intent and find the real home screen.

Also follow method refactoring.

Bug: 22358539
Change-Id: Iebc4ad8d2dd62ada079cab03d5765f7631fd4beb