Add a Developer option of "Disable automatic update".
For devices using A/B update or update-on-boot feature, we need a
Developer option to prevent the system updates from being installed
automatically.

We store the "disabled" status internally to turn on the automatic
update by default. But we show the opposite "enabled" status on screen
based on UX comments.

Bug: 27193001
Change-Id: I36a08a2a08fd1ba0f8f3c4b3ae5a08dc50829cd2