Don't show app in deletion helper if it was installed recently.
This solves several problems. The first is that if a user installs
a large app which triggers the deletion helper, the deletion helper
may clear out the recently installed large app to free up space.

The second is that it appears that the usage stats may not be
synced from device to device currently using Backup and Restore.
By considering the installation to be a use for deletion helper
filtering, this should avoid the problem for 60 days (and if the
user hasn't used the app at that point, it should be fair game
to clear).

Bug: 28800057
Change-Id: Idb8545aa45a42e45dc673dc5c179c0197204edfd