Materialize the button bar in the Deletion Helper.
Older Android style buttons were used for the bottom button bar.
By changing them to use the buttonBarBottomStyle, they are updated
to flat buttons. The bottom button bar for settings preference
fragments now more closely tracks the Material Design spec as seen
here in [0].

The other Settings views which use the button bar buttons are the
Redaction settings and WiFi network picker in some contexts. These
should be unaffected as they are based on the SettingsActivity,
rather than the SettingsFragment, and inflate a different view.

[0] https://material-design.storage.googleapis.com/publish/material_v_8/material_ext_publish/0B7WCemMG6e0VaDdZWXpLXzNicjg/components_buttons_usage6.png

Bug: 28845586
Change-Id: I9ea55aa2fe1e76f8f952880675501cbb0f76719d