Force dashboard summary to single-line and ellipsize it
Previously, the summary could become multiline. In rare combinations
of text and screen width, to indicate to the user that the text is
cut, we now ellipsize it too. Finally, since the ellipsis can get
too close to the end of the line, we add a padding equal to the left
margin of the icon.

Bug: 28872122
Change-Id: I27226e5de7781cf096508ddbef4f9095961ce028