Fix return when virtual keyboard setting changes.
The OnPreferenceChangeListener for the preference to show
the virtual keyboard was returning false, indicating that the
new preference was not to be persisted, even though it was
persisting the setting. That caused the accessibility
framework not to be informed of the new value, which led
TalkBack to speak the old value rather than the new one.

Bug: 30140972
Change-Id: I02b7dc1db52cb7646650e6f2e49a1e15c2233897