Get gestures animation preview image in async loader.
When user scrolls in Gestures Settings, the textureview used for the
animation video will be destroyed if the view is scrolled outside the
screen. When the view is scrolled back in, a slight flicker is seen as
the animation video is not instantly made available on the new surface
texture. We are now retrieving the preview image in an async loader
after the preference is created, so that it will not delay the original
loading of the gestures settings page, but the preview will be made
available while we initialize the animation video in the new surface
texture.

Bug: 29790087
Change-Id: I83f4a5bc4f24b7aab4c0cc851cb524c10e74f328