Check if InputMethodService is Direct Boot aware
This is a small follow up CL to the previous CL [1] that added
functionality to show a warning dialog when Direct Boot unaware apps and
IMEs are being selected.

In the previous CL, we checked whether the package to which the IME
belogs to is (fully or partially) Direct Boot aware.  If the package is
partially Direct Boot aware but the InputMethodService in question is
still Direct Boot unaware, the user will not see the warning dialog.

Luckily in InputMethodPreference we already have InputMethodInfo that
indirectly exposes ServiceInfo#directBootAware.  By directly checking
that bit we can simplify the logic and avoid such false negatives.

 [1]: I0498904d2f664fb41e8c1e6bb30d1cbf437cf4b9
      4a8136b51b746961aac6f5ce2459effef8c22680

Bug: 27196876
Change-Id: I869a7bd87748f09f7032a60b34ac0dbdc4a00b72