Fix crash when selecting AppInfo for uninstalled app from Recents.
When an app is uninstalled, the corresponding app info is not removed
fromi the Recents tasks. If user selects that app info entry from the
recents tasks, it will failed to retrieve the app detail as the package
is already removed.

Add the package removed event listener in InstalledAppDetails such
that when the package has been unintalled, its app info task will also
be removed from Recents.

Change-Id: I771507f8f53017f87eda31ddef77a95e30921cbf
Fixes: 28671923