Fix caption text being cut off when text size is set to very large.
In landscape mode, the caption preview height is set to 100dp. When
running in tablet and setting the caption text size to Very large, 100dp
is not enough to render the complete text with all the paddings defined.
Increasing the preview height to make enough room to display the caption
text.

Change-Id: Ia45c01a16eb1829d569d71648de398d48a7e4618
Fixes: 29125201