Stop apps with Storage Managers from launching activity when disabled
An app with a storage manager would be able to launch Manage Storage
from application settings, but when the app was disabled, the
activity would be null.

Test: In order to properly unit test this function, a large redesign is 
needed in this class, and is outside the scope of this bug.

Change-Id: Ia6b78bc6761bac8b701c904a6e438bbd951b63e8
Fixes: 30457616
(cherry picked from commit 8f17fcaa1dd85c816916ffbed3f66ad80c2bbf0a)