Fix accessibility title in confirm device credential fragment.
When setting the accessibility title in confirm credential fragment, it
simply does nothing if the title is null. But for most of the confirm
fragment only have header and details, not the title. In that case,
nothing will be set. Changing it to use the supplemental text directly
if the title is null.

Change-Id: Id9fdebf04535ccc8f38a933758a7948ee2966fb4
Fixes: 30221638