Fixed storage calculation when SD card is adopted as internal storage.
A SD Card can be set as internal storage, in which case its total must
be added to the total internal storage size.

Prior to this change, the total internal storage size was hardcoded as
the value returned by Storage Manager.

Also fixed the storage summary total to exclude portable devices.

Fixes: 30689428
BUG: 30742322

Change-Id: I1e33038eeebc91f3c525a387a97987391c10f2f9