Remove preference "Searching for Wi-Fi networks..."
When we add the preference "Searching for Wi-Fi networks...", we didn't
provide a key to the preference. As a result, each time when we get
access point update and try to show such tile, it creates a new one, and
results in multiple entries of such tile. Adding a key to the preference
will prevent the multiple entry from being added.

Test: Manual -
1. Turn wifi off and on several times, verify that there is only one entry
   of "Searching for Wi-Fi networks..." when initializing the wifi access
   point list.
2. Verifies that when the access point list is established, the
   "Searching for Wi-Fi networks..." entry is gone.

Change-Id: Iaad329d89f8e4b52129d0661501aa34fbd956692

Change-Id: I1bfa77051e068926e8c876a89adf14d143970e7e
Fixes: 31143961