Make a11y settings display consistent
We were using two methods to decide whether or not to show
that an a11y service was enabled, one on the main
Accessibility screen and another on the individual service
screen. In some situations, an enabled service would show
up as enabled on the main screen but not enabled on its
individual screen.

Bug: 30829589
Change-Id: I560d86684a2fb8987d3dd06279e033ecb7cc1a58