Fix bug preventing users from clicking settings suggestions twice
In an attempt to fix a monkey test crash which would open a
suggestion, click back and tap the same one again. The change
unfortunately prevented this for all users, but was not caught
because of a debug flag set on my device. A unit test will be added
to master to catch such cases, as tracked by this bug:
b/31937112

Change-Id: Ib1cff07e54903df687454bdbf2ed1416485ab34f
Fixes: 31926102
Test: Manual