Fix robolectric breakage because of internal resources
The robolectirc breakage is mainly because of ag/1726837 and the
current robo tests cannot deal with private resources. Robo engineer
will working on this to make sure robo tests is compatible with private
resources(b/33431346) and be globally used.

This is a temporary fix to stop the breakage by adding the hardcoded
color and will be reverted after b/33431346 is finished.

Test: make -j40 RunSettingsRoboTests
Change-Id: I5f451d94f0162c818b98aaaa49d452d926dd83b1