Remove the app header from ManageApplications when used for storage.
This UI element is no longer in the mocks. It complicated the
implementation of the fragment because it was causing the wrong
apps to be selected on-tap.

This specific StorageUseActivity usage only happens when coming
from the Storage Settings UI, so we can freely remove it without
harming other uses of ManageApplications.

Bug: 34195303
Test: Manual -- it seemed a bit silly to add a test to ensure that
      a broken UI element was indeed missing.

Change-Id: I5396364c297026a33b54d1794d9ad2c5102f0d7f