Fix a race condition where the extra view may not update timely.
Previously, we queried for the extra view's stats during the rebuild.
The full rebuild can take an arduous amount of time. By moving the
background calculation outside of the rebuild code path, we can
load the stats in the background and update the view outside of the
app loading cycle.

Bug: 35144044
Test: Manual -- this was race condition-y, so I just tested several
times.

Change-Id: Id2f92e32c3b3f9c5d5cf4a7308cca38c2e52789c