Update string: lockdown confirmation dialog title
The most common string shown previously was:
  "Set always-on VPN?"

It's now a different string:
  "Only allow connections through VPN?"

This makes more sense because the only way to get to this dialog in
stock Settings is by first enabling a specific VPN and afterwards
trying to set the same VPN to lockdown mode.

In this case the prompt should specifically refer to lockdown, not
having a VPN connection in general (because the user clearly already
has one).

Test: m Settings -j20, install a VPN, connect to it, open Settings > Network > Vpn > {VPN name} and select 'only allow connections through VPN'. Verify the title is now the one referred to in the commit message.
Fix: 34491069
Change-Id: Ic74a8b91eb8ae64fbe15723c173281f4ede52b06