Fix app filter breakage caused by ag/1900403
Bug: 35726115
Test: manual, run existing robo tests by make RunSettingsRoboTests

In ag/1900403 I renumbered the int constants in ManageApplications.java
to remove a gap in the sequence without realizing that the arrays
containing strings/filters based on these constants still contained
entries for the constant whose value was removed.

In this CL I've fixed that and reorganized the initialization code slightly
to make it easier to keep things in sync going forward.

Change-Id: I6d2db6499fb7eb080570bc757bde497fcd39fead