Specify disallowed action in AdminSupportDetailsDialog
The title of an AdminSupportDetailsDialog "Action not allowed" is
replaced by specialized strings such as "Screenshot not allowed" if
the action is specified by the extra DevicePolicyManager.EXTRA_RESTRICTION
in the calling intent.

This is done because these intents can be created by
third-party apps via the new public API
DevicePolicyManager.createAdminSupportIntent, and the disallowed action
might not be obvious in context of the third party app.

Bug: 31215663
Test: CtsVerifier in separate CL
Change-Id: I0c9e360cea34bec8eda29d86c5bf55d8c38c47ec