Disabling switch preference when unknown sources are blocked
When the user restriction DISALLOW_INSTALL_UNKNOWN_SOURCES is set,
changing the preferences in settings is meaningless as package installer
will not let the install proceed. Disabling the settings switch and
updated the summaries to reflect when a switch is disabled. Removing
robolectric test as the logic is no longer testable due to missing stubs
for UserManager and UserHandle.

Test: CtsVerifier -> Device Owner Tests -> Policy Transparency Test

Bug: 34688413
Change-Id: I7cf6cc09a5d29721186a0df9405d3c7e486295ee