Add progress bar and delay to WifiSettings refresh.
This change shows the progress bar for 0.3 seconds before AccessPoints are
about to be refreshed and 1.7 afterwards for a total duration of roughly
2 seconds. This indirectly fixes issues with no longer having a refresh button,
indicates ongoing scanning to the user, and should help avoid user errors where
the user clicks on the wrong access point while the list is being updated.

In order to avoid excess UI movement, the progress bar off state is now set
to invisible rather than gone.

Bug: b/34774783
Test: Currently blocked on refactor, tracked in b/36403635 and will
start next week.
Change-Id: I2bb6b5b3d4611cdbfd7138c758785601896d05b9