Remove ambiguous wildcard import of org.mockito.Matchers.*
The recent upgrade to Mockito 2.7.13 has caused some problems
for grok (and new javac toolchain). The changed files used
wildcard imports to import the same methods from both Matchers
and Mockito.

This was not a problem previously because in the previous
version of Mockito the Mockito class extended the Matchers class
and so even though two methods with the same name were imported
they were not treated as a conflict because they were treated as
being the same method. In 2.7.13 the Mockito class no longer
extends the Matchers class instead they both extend the
ArgumentMatchers class and so for some reason javac no longer
treats them as the same method.

This was not detected by presubmit builds because jack is
still treating the methods as being the same.

Bug: 32912773
Test: m -j32 ANDROID_FORCE_JACK_ENABLED=disabled FrameworksWifiTests
Change-Id: I8c2b1dd39d2d653fe0fba19ed6421b7de25bb82f