Fix crash in TetherSettings.
In onCreate(), it checkes whether ui is restricted, and try to show an
empty screen by creating a new preference screen. However, when the
preference is not attached, it failed to get the next preference id.
Instead of replacing the current screen with a new one, remove all the
preferences on the screen.

Change-Id: I085bee025cfeabaf4aefacb717d145e668bd85e1
Fix: 36589767
Test: manual