Fix build breakage in branches downstream of master.
R.string.background_activity_summary was deleted in master, but is
still referenced by res/xml/power_usage_details.xml. master builds
because the stale translations are currently still present, but will
suddenly fail to build as soon as the translations are updated.
Downstream branches of master which don't have translations configured
are already broken.

Reintroduce the old string to fix the builds.

Change-Id: I20c6aa41c97644a2733ff92a459587e3cb6bf7a0
Test: Will rely on TreeHugger
Fixes: 36619997