Don't enable a11y service if scary dialog obscured
We were only checking if the OK button itself was obscured.
Changing to check if any part of the window is obscured.

Bug: 31855267
Test: Created a sample a11y service that displays a UI to cover
part of the window, but not the OK button. It was possible to
enable an a11y service before this change, but now it displays
a warning toast.

Change-Id: I21974535aba734da6ea9d7accf67fb9237fd16fc